pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bunny-dns"
version = "0.1.0"
description = "Declaratively manage DNS zones, Pull Zones, and Edge Rules on bunny.net"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
bunny-dns = "bunny_dns.main:main"
bunny-dns-check = "bunny_dns.check_propagation:main"
